Move CachingAsyncSupplier
